Adding insert user answer

+++ service/user_service.py
@@ -4,8 +4,10 @@ from typing import Optional, List
 from api.internal_api.poll_service import user_answer_service_api
 from api.internal_api.poll_service.model.user_answer_response import UserAnswerResponse
 from model.user import User
+from model.user_answer_request import UserAnswerRequest
 
 from repository import user_repository
+from service import user_answer_service
 
 
 async def get_user_by_id(user_id: int) -> Optional[User]:
@@ -13,6 +15,7 @@ async def get_user_by_id(user_id: int) -> Optional[User]:
 
 async def create_user(user: User):
     await user_repository.create_user(user)
+    return {"message": "User create"}
 
 async def update_user(user_id: int, user: User):
     await user_repository.update_user(user_id,user)
@@ -22,14 +25,18 @@ async def register_user(user_id: int):
     return {"message": "User registered", "user_id": user_id}
 
 async def delete_user(user_id: int):
-    deleted_rows = await user_repository.delete_user(user_id)
-    return deleted_rows
-    ## ADD DELETING ALL ANSWERS FROM THE USER WITH USER_ID
+    user_exists = get_user_by_id(user_id)
+    if user_exists:
+        await user_repository.delete_user(user_id)
+        user_answer_response_list = await user_answer_service.get_user_answers_by_user_id(user_id)
+        if len(user_answer_response_list)==0:
+            return {"message": "User doesn't have any answered questions"}
+        else:
+            await user_answer_service_api.delete_user_answers(user_id)
+            return {"message":"User answers deleted"}
+    else:
+        return {"message": "User doesn't exist"}
 
-async def get_user_answers_by_user_id(user_id: int) -> Optional[List[UserAnswerResponse]]:
-    user_answer_response_list = await user_answer_service_api.get_user_answer(user_id)
-    return user_answer_response_list
-
-async def get_count_answers_by_user_id(user_id: int) -> int:
-    user_answered_count = await user_answer_service_api.get_count_answered_by_user(user_id)
-    return user_answered_count
+async def is_user_registered(user_id: int) -> bool:
+    is_registered = await user_repository.is_user_registered(user_id)
+    return is_registered
